[Refactor] Unify CLTV decode fn.
