+++ frontend/src/app/shared/script.utils.ts
@@ -556,3 +556,44 @@ export function isPoint(pointHex: string): boolean {
     return (y * y) % curveP === ySquared;
   }
 }
+
+export function detectCltvTimestamps(instructions: string[]): Map<number, number> {
+  const cltvTimestamps: Map<number, number> = new Map();
+  for (let i = 0; i < instructions.length; i++) {
+    const instruction = instructions[i];
+    const parts = instruction.split(' ');
+    const opcode = parts[0];
+    const args = parts.slice(1);
+
+    const pushMatch = opcode.match(/^PUSHBYTES_([1-4])$/);
+    if (pushMatch && args.length > 0) {
+      const byteCount = parseInt(pushMatch[1], 10);
+      const expectedLength = byteCount * 2;
+
+      if (args[0].length === expectedLength && /^[0-9a-fA-F]+$/.test(args[0])) {
+        if (i + 1 < instructions.length) {
+          const nextOpcode = instructions[i + 1].split(' ')[0];
+          if (nextOpcode === 'CLTV') {
+            const bytes = args[0].match(/.{2}/g) || [];
+            const littleEndianValue = bytes.reverse().join('');
+            const timestamp = parseInt(littleEndianValue, 16);
+            cltvTimestamps.set(i, timestamp);
+            cltvTimestamps.set(i + 1, timestamp);
+          }
+        }
+      }
+    }
+  }
+  return cltvTimestamps;
+}
+
+export function formatCltvTimestamp(timestamp: number): string {
+  if (timestamp < 500000000) {
+    return `Block height: ${timestamp}`;
+  }
+  if (timestamp > 4294967295) {
+    return `Invalid timestamp: ${timestamp}`;
+  }
+  const date = new Date(timestamp * 1000);
+  return date.toISOString().replace('T', ' ').replace(/\.\d{3}Z$/, ' UTC');
+}
